refactor(webapp): move PROD_ENVIRONMENT_NAME into utils/environments.ts
Extracted the constant from store/index.ts into a dedicated,
side-effect-free module so it can be imported in tests and other
non-store contexts without pulling in Zustand or QueryClient.

store/index.ts re-exports it for backwards compatibility.
