+++ packages/webapp/src/pages/Environment/Settings/General.tsx
@@ -9,7 +9,8 @@ import SettingsContent from './components/SettingsContent';
 import SettingsGroup from './components/SettingsGroup';
 import { useDeleteEnvironment, useEnvironment, usePatchEnvironment } from '../../../hooks/useEnvironment';
 import { useMeta } from '../../../hooks/useMeta';
-import { PROD_ENVIRONMENT_NAME, useStore } from '../../../store';
+import { useStore } from '../../../store';
+import { PROD_ENVIRONMENT_NAME } from '@/utils/environments';
 import { ConditionalTooltip } from '@/components-v2/ConditionalTooltip';
 import { EditableInput } from '@/components-v2/EditableInput';
 import { PermissionGate } from '@/components-v2/PermissionGate';
