+++ packages/webapp/src/pages/Root.tsx
@@ -3,7 +3,8 @@ import { useLocation, useNavigate } from 'react-router-dom';
 
 import { useMeta } from '../hooks/useMeta';
 import { useUser } from '../hooks/useUser';
-import { PROD_ENVIRONMENT_NAME, useStore } from '../store';
+import { useStore } from '../store';
+import { PROD_ENVIRONMENT_NAME } from '../utils/environments';
 
 /**
  * Rendered at `/`. Redirects the user to the right starting route based on
