+++ packages/webapp/src/utils/environments.ts
@@ -1,0 +1,2 @@
+/** The canonical name of the production environment. */
+export const PROD_ENVIRONMENT_NAME = 'prod';
